requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
pyyaml>=6.0
openai>=1.30.0
anthropic>=0.49.0
azure-identity>=1.16.0
gunicorn>=22.0.0
python-multipart>=0.0.9
